requirements.txt
# 渡の网关 - 依赖
flask>=3.0.0
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
# 农历（实时层按需注入「农历节气宜忌」用；无则仅返回占位）
zhdate>=0.1
